+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Add link to the course syllabus in an enrollment item
+  in the learner dashboard
+- Display Payment Schedule into SaleTunnel
 - Add OrderHelper allowEnrollment
 - Display Payment Schedule into SaleTunnel
 - Add order payment information in the dashboard, including a specific